requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
pillow==10.4.0
sqlparse==0.5.3
typing_extensions==4.13.2
